Change file structure and move files to appropriate folders
