+++ app/src/main/java/kapil/voiceassistedweatherapp/weather/WeatherDataProviderModule.java
@@ -1,4 +1,4 @@
-package kapil.voiceassistedweatherapp.modules;
+package kapil.voiceassistedweatherapp.weather;
 
 import android.content.Context;
 
@@ -6,7 +6,6 @@ import com.google.android.gms.common.api.GoogleApiClient;
 import com.google.android.gms.location.LocationServices;
 
 import javax.inject.Named;
-import javax.inject.Singleton;
 
 import dagger.Module;
 import dagger.Provides;
@@ -14,7 +13,7 @@ import retrofit2.Retrofit;
 import retrofit2.converter.jackson.JacksonConverterFactory;
 
 /**
- * Created by witworks on 18/04/17.
+ * This module is used to pass dependencies of {@link WeatherDataProvider}.
  */
 
 @Module
@@ -22,7 +21,6 @@ public class WeatherDataProviderModule {
 
     @Provides
     @Named("WIT_AI_SERVICE")
-    @Singleton
     Retrofit provideWitAiRetrofit() {
         return new Retrofit.Builder()
                 .baseUrl("https://api.wit.ai/")
@@ -32,7 +30,6 @@ public class WeatherDataProviderModule {
 
     @Provides
     @Named("WEATHER_SERVICE")
-    @Singleton
     Retrofit provideWeatherRetrofit() {
         return new Retrofit.Builder()
                 .baseUrl("http://api.openweathermap.org/")
@@ -41,7 +38,6 @@ public class WeatherDataProviderModule {
     }
 
     @Provides
-    @Singleton
     GoogleApiClient provideGoogleApiClient(Context context) {
         return new GoogleApiClient.Builder(context)
                 .addApi(LocationServices.API)
